test(a3p): refactor pushPrice script
rel: https://github.com/Agoric/BytePitchPartnerEng/issues/22
