+++ a3p-integration/proposals/f:replace-price-feeds/use.sh
@@ -5,5 +5,5 @@ set -e
 
 source /usr/src/upgrade-test-scripts/env_setup.sh
 
-./pushPrice.js 'ATOM' 12.01
-./pushPrice.js 'stATOM' 12.01
+./verifyPushedPrice.js 'ATOM' 12.01
+./verifyPushedPrice.js 'stATOM' 12.01
